services/knowledge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pyjwt==2.8.1
pydantic==2.5.0
numpy==1.24.3
httpx==0.25.2
sqlalchemy==2.0.23
python-multipart==0.0.6
